+++ HomeyPad/Utilities/Modulo.swift
@@ -1,4 +1,4 @@
-func modulo(_ a: Int8, _ n: Int8) -> Int8 {
+func modulo(_ a: Int, _ n: Int) -> Int {
     precondition(n > 0, "modulus must be positive")
     let r = a % n
     return r >= 0 ? r : r + n
